style: lint the code
Why should I prefer `Option::ok_or_else` instead of `Option::ok_or`?

The primary reason to use ok_or_else or any of the ..._or_else methods
is to avoid executing a function when it's not needed. In the case of
Option::ok_or_else or Option::unwrap_or_else, there's no need to run
extra code when the Option is Some. This can make code faster, depending
on what happens in the error case
refs: https://stackoverflow.com/a/45547401/13267147